refactor: moved useEffect hook for cart outside click to separate hook file
